+++ app/src/main/java/com/boilerplate/android/product_details/presentation/ProductDetailsViewState.kt
@@ -1,0 +1,9 @@
+package com.boilerplate.android.product_details.presentation
+
+import com.boilerplate.android.product_details.business.ProductDetails
+
+sealed class ProductDetailsViewState {
+    object Loading : ProductDetailsViewState()
+    data class Content(val product: List<ProductDetails>) : ProductDetailsViewState()
+    object Error : ProductDetailsViewState()
+}